Refactor main method Add Persistence Exception handling
Reason for this, and previous changes:
Instead of handling and propagating exception all the way to the top, now we're catching it in main. That's because we're using SQLite as database, and in case of any SQLException it means something deeply wrong with the environment where the app is running. So instead of trying to recover from that we're just closing the app